refactored artGallery controller/routes into submissions
