+++ server/index.js
@@ -28,7 +28,6 @@ require('./models/testModel');
 const testRouter = require('./routes/testRoutes');
 const submissionRouter = require('./routes/submissionRoutes');
 const userRouter = require('./routes/userRoutes');
-const artGalleryRouter = require('./routes/artGalleryRoutes');
 
 // Start the Node Express server
 const app = express();
@@ -40,7 +39,6 @@ app.use(express.urlencoded({ limit: '200mb', extended: true }));
 app.use('/test', testRouter);
 app.use('/submissions', submissionRouter);
 app.use('/user', userRouter);
-app.use('/artgallery', artGalleryRouter);
 
 app.get('/', (req, res) => {
   res.send('Hello World!');
