--- server/routes/artGalleryRoutes.js
@@ -1,10 +1,0 @@
-const express = require('express');
-
-const artGalleryRouter = express.Router();
-const artGalleryController = require('../controllers/artGalleryController');
-
-artGalleryRouter.get('/get', artGalleryController.getAllSubmissions);
-
-artGalleryRouter.get('/getsubmission', artGalleryController.getSubmission);
-
-module.exports = artGalleryRouter;
