+++ server/routes/submissionRoutes.js
@@ -5,7 +5,7 @@ const submissionController = require('../controllers/submissionController');
 
 submissionRouter.post('/post', submissionController.createSubmission);
 submissionRouter.delete('/delete/:id', submissionController.deleteSubmission);
-submissionRouter.get('/get', submissionController.getSubmissions);
-submissionRouter.get('/get/:id', submissionController.getSubmission);
+submissionRouter.get('/get', submissionController.getAllSubmissions);
+submissionRouter.get('/getsubmission', submissionController.getSubmission);
 
 module.exports = submissionRouter;
